# Cleanup type inference errors