--- a/src/DirectSearch.jl
+++ b/src/DirectSearch.jl
@@ -1,5 +1,6 @@
 module DirectSearch
 
+include("./Utils.jl")
 
 include("./Types.jl")
 include("./Cache.jl")