--- a/src/Poll.jl
+++ b/src/Poll.jl
@@ -17,7 +17,7 @@ Generate a set of directions with the configured polling algorithm, then return
 the set of points these directions give from the incumbent points.
 """
 function GeneratePollPoints(p::DSProblem{T}, ::AbstractMesh)::Vector{Vector{T}} where T
-    points = []
+    points = Vector{T}[]
     dirs = GenerateDirections(p)
 
     if !isnothing(p.x)