--- a/src/Types.jl
+++ b/src/Types.jl
@@ -85,22 +85,22 @@ mutable struct Config{T}
     end
 end
 
-mutable struct Status{T}
-    function_evaluations::Int
-    iteration::Int
+mutable struct Status
+    function_evaluations::Int64
+    iteration::Int64
     optimization_status::OptimizationStatus
 
     #= Time Running Totals =#
-    runtime_total::T
-    search_time_total::T
-    poll_time_total::T
-    blackbox_time_total::T
+    runtime_total::Float64
+    search_time_total::Float64
+    poll_time_total::Float64
+    blackbox_time_total::Float64
 
     #= Start/End Time =#
-    start_time::T
-    end_time::T
+    start_time::Float64
+    end_time::Float64
 
-    function Status{T}() where T
+    function Status()
         s = new()
 
         s.function_evaluations = 0