--- a/src/Utils.jl
+++ b/src/Utils.jl
@@ -1,0 +1,4 @@
+# Create an inf value in a given float type
+inf(::Type{Float64}) = Inf64
+inf(::Type{Float32}) = Inf32
+inf(::Type{Float16}) = Inf16