I have been going through the code using `@code_warntype` and we have had numerous places in the package where the compiler was unable to perform type inference at compile time leading to inefficient dispatching at runtime. This PR contains the high level changes to the main parts of the code, but there are still some type stability problems lurking in the depths of the code (such as the OrthoMADS computations). The main thing I have done is convert the abstract types inside of the structures to be concrete types and make the main problem then be parametric on them. There were also some problems using non-Float64 numbers (such as Float32), so I cleaned up the typing inside the constraints where that seemed to be the largest problem.

I benchmarked this on the rosenbrock function and here are the times I saw:

Before:
```julia
julia> @benchmark Optimize!(p) setup=(p = DSProblem(3); SetObjective( p, DS.rosenbrock ))
BenchmarkTools.Trial:
  memory estimate:  9.08 MiB
  allocs estimate:  167154
  --------------
  minimum time:     16.152 ms (0.00% GC)
  median time:      16.368 ms (0.00% GC)
  mean time:        17.082 ms (2.46% GC)
  maximum time:     27.947 ms (22.48% GC)
  --------------
  samples:          293
  evals/sample:     1
```

After:
```julia
julia> @benchmark Optimize!(p) setup=(p = DSProblem(3); SetObjective( p, DS.rosenbrock ))
BenchmarkTools.Trial:
  memory estimate:  7.08 MiB
  allocs estimate:  88178
  --------------
  minimum time:     11.699 ms (0.00% GC)
  median time:      11.802 ms (0.00% GC)
  mean time:        12.262 ms (2.65% GC)
  maximum time:     18.065 ms (22.72% GC)
  --------------
  samples:          408
  evals/sample:     1
```