@EdwardStables 2020-08-20
This is really great. 

You mentioned that you've changed types in Status to Float64 as that is the return type of time. Are we able to guarantee that this behavior is true on non 64-bit systems?