@imciner2 2020-08-20
> You mentioned that you've changed types in Status to Float64 as that is the return type of time. Are we able to guarantee that this behavior is true on non 64-bit systems?

Yep, it looks like we get a Float64 even on 32-bit systems. I looked into the base code that provides time here: https://github.com/JuliaLang/julia/blob/master/base/libc.jl#L249, and it is always casting to Float64 on return. So we should be fine with hardcoding those time variables to be always Float64.